refactor: change html lang
